Move Portal element to @clayui/shared
