+++ packages/clay-drop-down/package.json
@@ -30,6 +30,7 @@
 		"@clayui/checkbox": "^3.0.0",
 		"@clayui/icon": "^3.0.0",
 		"@clayui/radio-group": "^3.0.0",
+		"@clayui/shared": "^3.0.0",
 		"classnames": "^2.2.6",
 		"metal-position": "^2.1.2"
 	},
