+++ packages/clay-modal/package.json
@@ -22,6 +22,7 @@
 	"dependencies": {
 		"@clayui/button": "^3.0.0",
 		"@clayui/icon": "^3.0.0",
+		"@clayui/shared": "^3.0.0",
 		"classnames": "^2.2.6"
 	},
 	"peerDependencies": {
